pyproject.toml
[tool.poetry]
name = "sokolovd-29july2024"
version = "0.1.0"
description = ""
authors = ["danilasokolov0103gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
envyaml = "^1.10.211231"
allure-pytest = "^2.13.5"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
